event_scheduler/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
python-dateutil>=2.8.2
icalendar>=5.0.0
psycopg2-binary>=2.9.5
drf-yasg>=1.21.4
django-cors-headers>=3.13.0
python-dotenv>=0.21.0
pytest>=7.0.0
pytest-django>=4.5.2
flake8>=5.0.0
black>=22.10.0
isort>=5.10.1
uvicorn>=0.19.0
daphne>=4.0.0
channels>=4.0.0
channels-redis>=4.0.0
authlib>=1.2.0
jsonschema>=4.17.3
pytz>=2022.7.1